Move structured and imperative workflows to main `fixpoint` package
Move the structured and imperative workflow packages from
`fixpoint_extras` into the core `fixpoint` package.

There was an issue with circular imports after doing this. Basically, we
have an imperative workflow config which configures a Workflow and its
WorkflowRun resources. The config also configures memory, and in the
`fixpoint.memory` package, we let users pass in a `WorkflowRun` when
storing a memory so we can mark which workflow run the memory belongs
to.

This created a circular import loop.

To fix this, I added a `fixpoint._protocols.workflow_run` module and a
`WorkflowRunData` protocol, which lets us used structural typing. Now,
the `fixpoint.memory` uses `WorkflowRunData`, which overlaps with fields
in a `WorkflowRun`.

In a few places, I had to update references from `WorkflowRun` to
`WorkflowRunData`.
