+++ src/fixpoint/_protocols/workflow_run.py
@@ -1,0 +1,21 @@
+"""WorkflowRun protocol definition"""
+
+from typing import Protocol
+
+
+class WorkflowRunData(Protocol):
+    """A simple data container for workflow run info.
+
+    A simple data container for workflow run info. Technically, you could also
+    pass in `fixpoint.workflows.WorkflowRun`.
+
+    This class helps avoid circular imports.
+    """
+
+    @property
+    def id(self) -> str:
+        """The workflow run ID"""
+
+    @property
+    def workflow_id(self) -> str:
+        """The workflow ID"""
