+++ src/fixpoint/workflows/_workflow.py
@@ -1,26 +1,9 @@
-"""Simple implementation of a workflow"""
+"""The workflow and workflow run code.
 
-from dataclasses import dataclass
+A workflow is a collection of tasks and steps that are executed in a sequence.
+A workflow run is a running instance of a workflow.
+"""
 
-from .protocol import SupportsWorkflowRun
+__all__ = ["Workflow", "WorkflowRun"]
 
-
-@dataclass
-class WorkflowRun(SupportsWorkflowRun):
-    """A simple workflow run implementation"""
-
-    _id: str
-    _workflow_id: str
-
-    # pylint: disable=redefined-builtin
-    def __init__(self, id: str, workflow_id: str):
-        self._id = id
-        self._workflow_id = workflow_id
-
-    @property
-    def id(self) -> str:
-        return self._id
-
-    @property
-    def workflow_id(self) -> str:
-        return self._workflow_id
+from .imperative import Workflow, WorkflowRun
