+++ src/fixpoint/workflows/structured/_context.py
