+++ src/fixpoint/workflows/structured/_helpers.py
