+++ src/fixpoint/workflows/structured/errors.py
