+++ tests/workflows/structured/test_task.py
@@ -1,8 +1,7 @@
 import pytest
-import fixpoint
-from fixpoint_extras.workflows import imperative
-from fixpoint_extras.workflows import structured
-from fixpoint_extras.workflows.structured._run_config import RunConfig
+from fixpoint.workflows import imperative
+from fixpoint.workflows import structured
+from fixpoint.workflows.structured._run_config import RunConfig
 
 
 def test_task_declaration() -> None:
